candidate-evaluation/requirements.txt
# requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
supabase==2.0.2
openai>=1.13.3,<2.0.0
python-dotenv==1.0.0
langchain-openai==0.1.1